+++ src/components/Editable.tsx
@@ -67,6 +67,39 @@ const EMPTY_THOUGHT_TIMEOUT = 5 * 1000
 // eslint-disable-next-line jsdoc/require-jsdoc
 const stopPropagation = (e: React.MouseEvent) => e.stopPropagation()
 
+/** Add position:absolute to toolbar elements in order to fix Safari position:fixed browser behavior when keyboard is up. */
+const makeToolbarPositionFixed = () => {
+  const hamburgerMenu = document.getElementsByClassName('hamburger-menu')[0] as HTMLElement
+  const toolbar = document.getElementsByClassName('toolbar-container')[0] as HTMLElement
+  const rightArrow = document.getElementById('right-arrow') as HTMLElement
+  const leftArrow = document.getElementById('left-arrow') as HTMLElement
+  Array.from([hamburgerMenu, toolbar, rightArrow, leftArrow]).forEach(el => {
+    el.style.position = 'absolute'
+    if (el !== rightArrow && el !== leftArrow) {
+      el.style.top = `${window.scrollY}px`
+    }
+  })
+}
+/** Reset position:absolute of toolbar elements. */
+const resetToolbarPosition = () => {
+  const hamburgerMenu = document.getElementsByClassName('hamburger-menu')[0] as HTMLElement
+  const toolbar = document.getElementsByClassName('toolbar-container')[0] as HTMLElement
+  const rightArrow = document.getElementById('right-arrow') as HTMLElement
+  const leftArrow = document.getElementById('left-arrow') as HTMLElement
+  Array.from([hamburgerMenu, toolbar, rightArrow, leftArrow]).forEach(el => {
+    el.style.position = 'fixed'
+    el.style.top = ''
+  })
+}
+/** Update position of toolbar elements while scrolling in order to show them always on top. */
+const updateToolbarPositionOnScroll = () => {
+  const hamburgerMenu = document.getElementsByClassName('hamburger-menu')[0] as HTMLElement
+  const toolbar = document.getElementsByClassName('toolbar-container')[0] as HTMLElement
+  Array.from([hamburgerMenu, toolbar]).forEach(el => {
+    el.style.top = `${window.scrollY}px`
+  })
+}
+
 interface EditableProps {
   contextChain: Child[][],
   cursorOffset?: number,
@@ -352,6 +385,9 @@ const Editable = ({ disabled, isEditing, thoughtsRanked, contextChain, cursorOff
 
   /** Flushes edits and updates certain state variables on blur. */
   const onBlur = () => {
+    resetToolbarPosition()
+    document.removeEventListener('scroll', updateToolbarPositionOnScroll)
+
     const { invalidState } = state
     throttledChangeRef.current.flush()
 
@@ -377,6 +413,8 @@ const Editable = ({ disabled, isEditing, thoughtsRanked, contextChain, cursorOff
    * Prevented by mousedown event above for hidden thoughts.
    */
   const onFocus = () => {
+    makeToolbarPositionFixed()
+    document.addEventListener('scroll', updateToolbarPositionOnScroll)
 
     // must get new state
     const state = store.getState()
